RD-1223 Merge Page container with component
